requirements.txt
PySide6==6.7.2
PySide6-Addons==6.7.2
markdown-it-py==3.0.0
mdurl==0.1.2

